+++ apps/web/src/Providers.tsx
@@ -4,8 +4,9 @@ import { Store } from '@reduxjs/toolkit'
 import { HydrationBoundary, QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { HistoryManagerProvider } from 'contexts/HistoryContext'
 import { ThemeProvider as NextThemeProvider, useTheme as useNextTheme } from 'next-themes'
+import { useMemo } from 'react'
 import { Provider } from 'react-redux'
-import { wagmiConfig } from 'utils/wagmi'
+import { createWagmiConfig } from 'utils/wagmi'
 import { WagmiProvider } from 'wagmi'
 
 // Create a client
@@ -23,6 +24,7 @@ const StyledUIKitProvider: React.FC<React.PropsWithChildren> = ({ children, ...p
 const Providers: React.FC<
   React.PropsWithChildren<{ store: Store; children: React.ReactNode; dehydratedState: any }>
 > = ({ children, store, dehydratedState }) => {
+  const wagmiConfig = useMemo(() => createWagmiConfig(), [])
   return (
     <WagmiProvider reconnectOnMount config={wagmiConfig}>
       <QueryClientProvider client={queryClient}>
