+++ apps/web/src/utils/viem.ts
@@ -1,7 +1,10 @@
 import { ChainId } from '@pancakeswap/chains'
+import first from 'lodash/first'
+import { PublicClient, createPublicClient, fallback, http } from 'viem'
+import { mainnet } from 'viem/chains'
+
 import { CHAINS } from 'config/chains'
 import { PUBLIC_NODES } from 'config/nodes'
-import { PublicClient, createPublicClient, fallback, http } from 'viem'
 
 export type CreatePublicClientParams = {
   transportSignal?: AbortSignal
@@ -55,4 +58,33 @@ export function createViemPublicClientGetter({
   return function getClients({ chainId }: { chainId?: ChainId }): PublicClient {
     return clients[chainId as ChainId]
   }
+}
+
+const PUBLIC_MAINNET = 'https://ethereum.publicnode.com'
+
+export const CLIENT_CONFIG = {
+  batch: {
+    multicall: {
+      batchSize: 1024 * 200,
+      wait: 16,
+    },
+  },
+  pollingInterval: 6_000,
+}
+
+export const publicClient = ({ chainId }: { chainId?: ChainId }) => {
+  if (chainId && viemClients[chainId]) {
+    return viemClients[chainId]
+  }
+  let httpString: string | undefined
+
+  if (process.env.NODE_ENV === 'test' && chainId === mainnet.id) {
+    httpString = PUBLIC_MAINNET
+  } else {
+    httpString = chainId && first(PUBLIC_NODES[chainId]) ? first(PUBLIC_NODES[chainId]) : undefined
+  }
+
+  const chain = CHAINS.find((c) => c.id === chainId)
+
+  return createPublicClient({ chain, transport: http(httpString), ...CLIENT_CONFIG })
 }
